Refactor JenkinsService (#3570)
